Unified structure of main and mods

+++ MOD_RIEMANN/stls.h
@@ -7,35 +7,35 @@
 // FUNCTIONS USED TO ALLOCATE AND FREE ARRAYS
 // -------------------------------------------------------------------
 
-void alloc_stls_arrays(input in, float **xx, float **phi,
-		       float **GG, float **GG_new,
-		       float **SS, float **SSHF);
+void alloc_stls_arrays(input in, double **xx, double **phi,
+		       double **GG, double **GG_new,
+		       double **SS, double **SSHF);
 
-void free_stls_arrays(float *xx, float *phi, float *GG,
-                      float *GG_new, float *SS,
-                      float *SSHF);
+void free_stls_arrays(double *xx, double *phi, double *GG,
+                      double *GG_new, double *SS,
+                      double *SSHF);
 
 
 // -------------------------------------------------------------------
 // FUNCTION USED TO INITIALIZE ARRAYS
 // -------------------------------------------------------------------
 
-void init_fixed_stls_arrays(input *in, float *xx, 
-			    float *phi, float *SSHF, bool verbose);
+void init_fixed_stls_arrays(input *in, double *xx, 
+			    double *phi, double *SSHF, bool verbose);
 
 // -------------------------------------------------------------------
 // FUNCTION USED TO COMPUTE THE CHEMICAL POTENTIAL
 // -------------------------------------------------------------------
 
-float compute_mu(input in);
+double compute_mu(input in);
 
-float normalization_condition(float mu, void *pp);
+double normalization_condition(double mu, void *pp);
 
 // ------------------------------------------------------------------
 // FUNCTION USED TO DEFINE THE WAVE-VECTOR GRID
 // ------------------------------------------------------------------
 
-void wave_vector_grid(float *xx, input in);
+void wave_vector_grid(double *xx, input in);
 
 
 // -------------------------------------------------------------------
@@ -49,50 +49,50 @@ int idx2(int xx, int yy, int x_size);
 // FUNCTIONS USED TO COMPUTE THE NORMALIZED IDEAL LINDHARD DENSITY
 // -------------------------------------------------------------------
 
-void compute_phi(float *phi, float *xx, input in, bool verbose);
+void compute_phi(double *phi, double *xx, input in, bool verbose);
 
-void compute_phil(float *phil, float *xx, int ll, input in);
+void compute_phil(double *phil, double *xx, int ll, input in);
 
-float phixl(float yy, float xx, int ll, input in);
+double phixl(double yy, double xx, int ll, input in);
 
-float phix0(float yy, float xx, input in);
+double phix0(double yy, double xx, input in);
 
 // -------------------------------------------------------------------
 // FUNCTIONS USED TO COMPUTE THE STATIC STRUCTURE FACTOR
 // -------------------------------------------------------------------
 
-void compute_ssf(float *SS, float *SSHF,
-                 float *GG, float *phi, 
-		 float *xx, input in);
+void compute_ssf(double *SS, double *SSHF,
+                 double *GG, double *phi, 
+		 double *xx, input in);
 
-void compute_ssfHF(float *SS,  float *xx, input in);
+void compute_ssfHF(double *SS,  double *xx, input in);
 
-float ssfHF(float yy, float xx, input in);
+double ssfHF(double yy, double xx, input in);
 
 // -------------------------------------------------------------------
 // FUNCTIONS USED TO COMPUTE THE STATIC LOCAL FIELD CORRECTION
 // -------------------------------------------------------------------
 
-void compute_slfc(float *GG, float *SS, 
-		  float *xx, input in);
+void compute_slfc(double *GG, double *SS, 
+		  double *xx, input in);
 
-float slfc(float yy, float xx, float SS);
+double slfc(double yy, double xx, double SS);
 
 // -------------------------------------------------------------------
 // FUNCTIONS USED TO COMPUTE THE INTERNAL ENERGY
 // -------------------------------------------------------------------
 
-float compute_uex(float *SS, input in);
+double compute_uex(double *SS, input in);
 
 // -------------------------------------------------------------------
 // FUNCTIONS FOR OUTPUT AND INPUT
 // -------------------------------------------------------------------
 
-void write_text(float *SS, float *GG, float *phi, 
-		float *SSHF, float *xx, input in);
+void write_text(double *SS, double *GG, double *phi, 
+		double *SSHF, double *xx, input in);
 
-void write_guess(float *SS, float *GG, input in);
+void write_guess(double *SS, double *GG, input in);
 
-void read_guess(float *SS, float *GG, input in);
+void read_guess(double *SS, double *GG, input in);
 
 #endif
